update : change dtype (bfloat16 -> float16)

+++ llm/qwen3/config.py
@@ -13,5 +13,5 @@ class QwenConfig:
     hidden_dim : int = 3072
     qk_norm : bool = True
     rope_base : float = 1e6
-    dtype : torch.dtype = torch.bfloat16
+    dtype : torch.dtype = torch.float16
